Remove Gradle propdeps plugin
This commit removes the propdeps plugin.
"provided" configurations are replaced with "compileOnly", "testCompile"
configurations.
The "optional" configuration supported by propdeps is now replaced by a
local Gradle plugin defining a specific configuration for that.

As a result of that change, optional/provided dependencies are not
published with the POMs anymore.