Adopt new Gradle maven-publish plugin
The maven plugin was removed in Gradle 7, so we need to replace it with
maven-publish. This uses a similar DSL, but operates in a slightly
different way. The simplest way of handling the differences is to
reorder its application within build.gradle so that appropriate
properties will have been applied by the time publish-maven.gradle
is invoked.