Adopt Spring Framework 6 and Jakarta EE 9
To avoid any complications with functional changes between Hibernate 5
and Hibernate 6, this uses the hibernate-core-jakarta v5 module rather
than hibernate-core v6. This requires a specific dependency exclusion
in spring-webflow.gradle to make sure that hibernate-entitymanager
never causes both versions to be loaded. This can be simplified after
eventual adoption of Hibernate 6.

The Spring snapshot repository needs to be referenced in build.gradle
for this early stage work on SWF 3.